Worked on User Authentication and Middleware and Data Fetching
Created a component that fetches the users name via server side. Tried to put that inside the layout for the dashboard. Implemented middleware for all pages except home page. Tried to work on redirecting the user to the home page after logging out and and the home page would refresh if you log out but so far unsuccessful. Should look into callbacks for signOut via nextAuth.